Removed the custom events handler
We were registering events in details files to be fired by the cores events handler, this is no longer needed as it was caused by an overlooked error in the way we handled naming of modules. This commit doesn't affect the core much (if at all), but it does require additional modules to be updated to reflect these new changes and use their own events file - as it should be.