Category query problem
I'm trying to understand why its happening and what i'm seeing...  
On the backend my four products all have the correct category in the edit and index screen of products.  But on the front end that category shows only 2 of the products in the category??? all products are Live, as is the category.  

Digging around the select statement ( build_query in categories ) seems to be flawed .. ( maybe im wrong ), there isn't a reason to use a second join to pull the category information when your only retrieving product ids. 