Added uncategorized string to the lang files.
In this way the user can have the "Uncategorized" category translated
in his own language.
Before the uncategorized string were the same for all the languages,
now for exampe in Italian I can see the correct translation for my
language "Senza categoria" and this will be much more usefull for
costumers that will use the module in different language.